command/format: maintain correspondance for list diffs
Previously we were trying to detect and show insertions and deletions in
the middle of the list as such, rather than showing them as cascading
updates of all following items.

Unfortunately, while this result looks nicer in the diff rendering it is
rather misleading when considering how Terraform actually treats list
changes internally, and so it's better to represent the one-to-one
index-based correspondence to illustrate that Terraform will see
insertions as cascading updates to the end of the list because that
better explains the resulting provider behavior, particularly when
the nested block contains Computed attributes that need to be preserved
on later updates.